pallet_chain_manager::pallet

Module dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

chill

Request to step back or “chill” immediately from election participation by erasing affidavit keys.

declare

Submit an affidavit for the upcoming session election.

elect

Execute the election for the upcoming session.

force_genesis_config

Force-update a selected genesis configuration parameter.

inspect_affidavit

Emit the stored affidavit for a given affidavit identifier as an event.

inspect_elects

Emit the currently revealed elected author set as an event for inspection.

prepare_validation_payload

Construct and emit a signed ValidatePayload and its signature as an event.

validate

Register an author’s new affidavit signing key for the upcoming session’s participation as a validator.