pallet_chain_manager::pallet::dispatchables

Function validate

pub fn validate<T: Config>(
    payload: ValidatePayloadOf<T>,
    _signature: T::Signature,
)

Register an author’s new affidavit signing key for the upcoming session’s participation as a validator.

This extrinsic allows an author to declare their off-chain signing key that will be used to sign affidavits for the next session’s election.

It ensures that only valid and available authors can register keys.

Notes

Affidavit Keys are session-specific and are updated each session.

This extrinsic allows submission of affidavit for next session only.

Affidavit Declaration/Submission will rotate keys once every submission for further sessions then.

Errors

Returns a DispatchError if author-role authorization fails.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::validate.