pallet_chain_manager::pallet

Type Alias AffidavitKeys

pub type AffidavitKeys<T: Config> = StorageNMap<_GeneratedPrefixForStorageAffidavitKeys<T>, (NMapKey<Blake2_128Concat, SessionIndex>, NMapKey<Blake2_128Concat, AffidavitId<T>>), AuthorOf<T>, OptionQuery>;

Offchain affidavit keys registered by an author for a given session.

Retrieves the author who submitted affidavit keys, tied to a specific session.

Authenticate offchain submissions (affidavits) for the upcoming session.

Ensure that only valid authors can participate in elections or submit weights.

Maintain session-specific separation, so keys do not persist beyond their intended session.

Storage Structure

SessionIndex: the session for which the keys are valid.

AffidavitId: the actual affidavit public key ID, type chosen by the runtime configuration Config::AffidavitCrypto.

AuthorOf<T>: the author who owns the key-pairs for signing.

Storage type is [StorageNMap] with keys type (SessionIndex, AffidavitId < T >) and value type AuthorOf < T >.

Aliased Type

pub struct AffidavitKeys<T: Config>(/* private fields */);