pallet_chain_manager::pallet

Type Alias ElectionRunnerPointsUpgrade

pub type ElectionRunnerPointsUpgrade<T: Config> = StorageValue<_GeneratedPrefixForStorageElectionRunnerPointsUpgrade<T>, Option<T::Points>, ValueQuery>;

Pending upgrade of election runner points.

When an upgrade occurs, this should be set to None after being applied.

Notes

Upgrades should occur after all author rewards have been distributed and before the next reward cycle starts.

Ensures fairness: if upgraded mid-session, authors who have not yet submitted affidavits may be disadvantaged in election chances.

Any user of ElectionRunnerPoints should take this storage value into account for the upcoming session.

Storage type is [StorageValue] with value type Option < T :: Points >.

Aliased Type

pub struct ElectionRunnerPointsUpgrade<T: Config>(/* private fields */);