pallet_chain_manager

Module pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

The main components of this pallet are:

Pallet, which implements all of the dispatchable extrinsics of the pallet, among other public functions.

The subset of the functions that are dispatchable can be identified either in the dispatchables module or in the Call enum.

storage_types, which contains the list of all types that are representing a storage item. Otherwise, all storage items are listed among Type Definitions.

Config, which contains the configuration trait of this pallet.

Event and Error, which are listed among the Enums.

Modules

dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Structs

GenesisConfig

Genesis configuration for the Chain Manager pallet.

Internals 🔒

Internal helper struct for implementing not-exposable blockchain trait operations.

Pallet

Primary Marker type for the Chain Manager pallet.

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Error

The Error enum of this pallet.

Event

The Event enum of this pallet

Traits

Config

Configuration trait of this pallet.

Type Aliases

AffidavitBeginsAt

Start of affidavit submission window as a percentage ([PerThing]) of the current session. Authors submit affidavits for the next upcoming session starting from this point.

AffidavitEndsAt

End of affidavit submission window as a percentage ([PerThing]) of the current session.

AffidavitKeys

Offchain affidavit keys registered by an author for a given session.

AffidavitTxPriority

Transaction priority for the Pallet::declare extrinsic.

AllowAffidavits

Flag to enable or disable affidavit submissions.

AuthorAffidavits

Storage mapping of authors’ submitted affidavits per session.

BlockPointsStore

Author points accumulated per session.

CurrentSession

The current running session index.

ElectionBeginsAt

Start of the election window as a percentage ([PerThing]) of the affidavit submission period in the current session.

ElectionRunnerPoints

Block points allocated for the election runner.

ElectionRunnerPointsUpgrade

Pending upgrade of election runner points.

ElectionTxPriority

Transaction priority for the Pallet::elect extrinsic.

ElectsPreparedBy

Tracks the author who prepared the election for a given session.

FinalityAfter

Wall-clock delay (in timestamp units) that must elapse after the first observation of a value before it becomes eligible to strengthen its confidence signal.

FinalityTicks

Number of distinct block observations required after the FinalityAfter window has elapsed in order to strengthen the confidence signal of a value.

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

SessionStartAt

Start block of the current session.

ValidateTxPriority

Transaction priority for the Pallet::validate extrinsic.