pallet_chain_manager::pallet::dispatchables

Function chill

pub fn chill<T: Config>(affidavit_pub: AffidavitId<T>)

Request to step back or “chill” immediately from election participation by erasing affidavit keys.

This safely avoids penalties by pausing an author’s duties.

This extrinsic allows an author to withdraw from participating in upcoming elections or prevent submitting future affidavits. The actual effect depends on the current block relative to the affidavit submissions and election windows.

Note that its always advised to chill validation before resign author role to skip unnecessary invalid affidavit declarations.

Parameters

origin: Must be a signed account corresponding to the author i.e., controller/role account.

affidavit_pub : Public affidavit key registered for a new session’s affidavit submission.

Notes

Removing affidavit keys ensures authors cannot unfairly influence future elections.

By inspecting returned errors, callers can compute the optimal chill window.

Errors

Returns a DispatchError with diagnostic in case of irrevocable duties assigned.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::chill.